Updated protocol definitions: - revamped the event definitions. Now events are split in two categories: rider events and rider events. Each time of an event has its own message - Added `updateCurrentLocation` to `UserService` - Removed `getPendingRides`, replaced it with  `subscribeForRideRequests`
Introduced `UserRegistry` class that is responsible for maintaining in-memory user and driver temporary representations which have their sessionId, streamObserver (to interact with a client). Also,  `UserRegistry` provides the closest available driver.
Added tests for `UserRegistry`

Introduced a state machine that will execute the ride flow based on the exchange of events between the rider, the driver and the system.

Signed-off-by: Elizaveta Sorokina