Introduced `RouteServer` with two implementations. Mock implementation generates a straight line with intermediate points for testing. `GoogleMapsRouteService` is a real one but it's not properly configured yet
Refactored `RideStateMachineConfig` - moved out logic to appropriate services to keep the config readable and compact. Added `DriverService` to host that logic.

Removed redundant states from the state machine.

Implemented better representation of a route in the protocol.

Added `driverArrived` method to `DriverService`

Added `DriverLocationUpdateService` which periodically updates rider with the latest driver's location

Added a test simulating the driver flow up to a point of driver's arrival for a pickup

Signed-off-by: Elizaveta Sorokina